% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2consensus.R
\name{bam2consensus}
\alias{bam2consensus}
\title{bam2consensus}
\usage{
bam2consensus(dna, x.pos = NULL, x.name = "x", chr.name = "chr",
  wlen = 25000, wjump = 25000, start.by = 1, end.by = NULL,
  wtype = "bp", dist = "IUPAC", global.deletion = TRUE,
  threads = 1, ncores = 1, pB = TRUE)
}
\arguments{
\item{chr.name}{chromosome name [default: NULL]}

\item{wlen}{sliding windows length [default: 100000]}

\item{wjump}{sliding windows jump [default: 100000]}

\item{start.by}{optional start position [default: NULL]}

\item{end.by}{optional end position [default: NULL]}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites}
or \code{triSites} [default: "bp"] see \link[distIUPAC]{biSites} and
\link[distIUPAC]{triSites}}

\item{dist}{distance to use}

\item{global.deletion}{a logical indicating whether to delete the sites
with missing data in a global [TRUE] or pairwise way [FALSE]}

\item{threads}{number of parallel threads to process sliding windows}

\item{ncores}{number of parallel cores to process pairwise
distance calculation [default: 1] see \link[distIUPAC]{rcpp_distIUPAC}}

\item{pB}{specifies if progress should be shown as a progress bar}

\item{pop.list}{population list}

\item{pop.names}{population names [default: NULL]}
}
\description{
This function calculates \code{distIUPAC} based distances
within one population (x: receiver; x: donor) along sliding window.
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
##multiple threads to process 
CAS.xStats<-xStats(MySequences, x.pos=CAS.pos, x.name="CAS", threads=2)
CAS.xStats
}
\author{
Kristian K Ullrich
}
