% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.xyioStats.R
\name{dist.xyioStats}
\alias{dist.xyioStats}
\title{dist.xyioStats}
\usage{
dist.xyioStats(tmpSEQ, x.pos, y.pos, i.pos, o.pos, x.name = "x",
  y.name = "y", i.name = "i", o.name = "o", dist = "IUPAC",
  ncores = 1)
}
\arguments{
\item{tmpSEQ}{\code{DNAStringSet} [mandatory]}

\item{x.pos}{population X positions [mandatory]}

\item{y.pos}{population Y positions [mandatory]}

\item{i.pos}{population I positions [mandatory]}

\item{o.pos}{population O positions [mandatory]}

\item{x.name}{population X name [default: "x"]}

\item{y.name}{population Y name [default: "y"]}

\item{i.name}{population I name [default: "i"]}

\item{o.name}{population I name [default: "o"]}

\item{dist}{distance to use, choose one model as in
\link[ape]{dist.dna} or [default: "IUPAC"]}

\item{ncores}{number of parallel cores to process pairwise distance
calculation [default: 1]}
}
\description{
This function calculates \code{distIUPAC} based distances
comparing two populations (x: receiver; y: donor) with an ingroup
population (i: ingroup) and an outgroup population (o: outgroup).
In a four-taxon scenario (((P1,P2),P3),O) with geneflow from P3>>P2,
the populations should be defined for deltaMean and deltaMin statistics
as follows [x:P2 y:P3 i:P1 o:P4].
Accordingly in the four-taxon scenario (((P1,P2),P3),O) with geneflow from
P2>>P3, the populations should be defined for deltaMean and deltaMin
statistics as follows [x:P3 y:P2 i:P1 o:P4].
}
\examples{
##Use the 'xyioStats' function to handle population assignment automatically

##load sequence data
data("MySequences", package="distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
IRA.pos<-71:78
APO.pos<-1

##Here, one needs to consider the changed x, y and i positions due
##to sub-sampling.
CAS.pos_<-seq_along(CAS.pos)
AFG.pos_<-seq(from=length(CAS.pos_)+1, to=length(c(CAS.pos, AFG.pos)))
IRA.pos_<-seq(from=length(c(CAS.pos, AFG.pos))+1,
to=length(c(CAS.pos, AFG.pos, IRA.pos)))
APO.pos_<-seq(from=length(c(CAS.pos, AFG.pos, IRA.pos))+1,
to=length(c(CAS.pos, AFG.pos, IRA.pos, APO.pos)))

##pairwise deletion
CAS_AFG_IRA_APO.pairwiseDeletion.dist<-distIUPAC(
  as.character(MySequences[c(CAS.pos,AFG.pos,IRA.pos,APO.pos)]))
CAS_AFG_IRA_APO.pairwiseDeletion.xyioStats<-dist.xyioStats(
  CAS_AFG_IRA_APO.pairwiseDeletion.dist, 
  x.pos=CAS.pos_, y.pos=AFG.pos_, i.pos=IRA.pos_, o.pos=APO.pos_,
  x.name="CAS", y.name="AFG", i.name="IRA", o.name="APO")

##global deletion
CAS_AFG_IRA_APO.globalDeletion.dist<-distIUPAC(
  as.character(globalDeletion(MySequences[c(CAS.pos,AFG.pos,
  IRA.pos,APO.pos)])))
CAS_AFG_IRA_APO.globalDeletion.xyioStats<-dist.xyioStats(
  CAS_AFG_IRA_APO.globalDeletion.dist,
  x.pos=CAS.pos_, y.pos=AFG.pos_, i.pos=IRA.pos_, o.pos=APO.pos_,
  x.name="CAS", y.name="AFG", i.name="IRA", o.name="APO")

##compare results
rbind(CAS_AFG_IRA_APO.pairwiseDeletion.xyiStats,
  CAS_AFG_IRA_APO.globalDeletion.xyiStats)
}
\references{
Slatkin, M. (1991). Inbreeding coefficients and coalescence 
times. \emph{Genetics Research}, \bold{58(2)}, 167-175.

Beerli, P. (1998). Structured Populations. \emph{Advances in molecular
ecology}, \bold{306}, 39.

Peter, B. M. (2016). Admixture, population structure, and F-statistics.
\emph{Genetics}, \bold{202(4)}, 1485-1501.

Paradis, E., & Schliep, K. (2018). ape 5.0: an environment for modern
phylogenetics and evolutionary analyses in R. \emph{Bioinformatics},
\bold{35(3)}, 526-528.

Rosenzweig, B. K., Pease, J. B., Besansky, N. J., & Hahn, M. W. (2016).
Powerful methods for detecting introgressed regions from population genomic
data. \emph{Molecular ecology}, \bold{25(11)}, 2387-2397.

Hahn, M. W., Hibbins, M. S. (2019). A Three-Sample Test for Introgression.
\emph{Molecular biology and evolution}, \bold{msz178}.
}
\seealso{
\code{\link[distIUPAC]{xyioStats}},
\code{\link[distIUPAC]{distIUPAC}}, \code{\link[distIUPAC]{rcpp_distIUPAC}},
\code{\link[ape]{dist.dna}}
}
\author{
Kristian K Ullrich
}
