% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.xyiStats.R
\name{dist.xyiStats}
\alias{dist.xyiStats}
\title{dist.xyiStats}
\usage{
dist.xyiStats(tmpSEQ, x.pos, y.pos, i.pos, x.name = "x", y.name = "y",
  i.name = "i", dist = "IUPAC", ncores = 1)
}
\arguments{
\item{tmpSEQ}{\code{DNAStringSet} [mandatory]}

\item{x.pos}{population X positions [mandatory]}

\item{y.pos}{population Y positions [mandatory]}

\item{i.pos}{population I positions [mandatory]}

\item{x.name}{population X name [default: "x"]}

\item{y.name}{population Y name [default: "y"]}

\item{i.name}{population I name [default: "i"]}

\item{dist}{distance to use [default: IUPAC] or choose one model as in
\link[ape]{dist.dna} [default: "IUPAC"]}

\item{ncores}{number of parallel cores to process pairwise distance
calculation [default: 1] see \link[distIUPAC]{rcpp_distIUPAC} [default: 1]}
}
\description{
This function calculates \code{distIUPAC} based distances
comparing two populations (x: receiver; y: donor) with an ingroup
population (i: ingroup).
In a four-taxin scenario (((P1,P2),P3),O) with geneflow from P3>>P2,
the populations should be defined for deltaMean and deltaMin statistics
as follows [x:P2 y:P3 i:P1].
Accordingly in the four-taxin scenario (((P1,P2),P3),O) with geneflow from
P2>>P3, the populations should be defined for deltaMean and deltaMin
statistics as follows [x:P3 y:P2 i:P1].
}
\examples{
##Use the 'xyiStats' function to handle population assignment automatically

##load sequence data
data("MySequences", package="distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
IRA.pos<-71:78

##Here, one needs to consider the changed x, y and i positions due
##to sub-sampling.
CAS.pos_<-seq_along(CAS.pos)
AFG.pos_<-seq(from=length(CAS.pos_)+1, to=length(c(CAS.pos, AFG.pos)))
IRA.pos_<-seq(from=length(c(CAS.pos, AFG.pos))+1,
to=length(c(CAS.pos, AFG.pos, IRA.pos)))

##pairwise deletion
CAS_AFG_IRA.pairwiseDeletion.dist<-distIUPAC(
  as.character(MySequences[c(CAS.pos,AFG.pos,IRA.pos)]))
CAS_AFG_IRA.pairwiseDeletion.xyiStats<-dist.xyiStats(
  CAS_AFG_IRA.pairwiseDeletion.dist, 
  x.pos=CAS.pos_, y.pos=AFG.pos_, i.pos=IRA.pos_,
  x.name="CAS", y.name="AFG", i.name="IRA")

##global deletion
CAS_AFG_IRA.globalDeletion.dist<-distIUPAC(
  as.character(globalDeletion(MySequences[c(CAS.pos,AFG.pos,IRA.pos)])))
CAS_AFG_IRA.globalDeletion.xyiStats<-dist.xyiStats(
  CAS_AFG_IRA.globalDeletion.dist,
  x.pos=CAS.pos_, y.pos=AFG.pos_, i.pos=IRA.pos_,
  x.name="CAS", y.name="AFG", i.name="IRA")

##compare results
rbind(CAS_AFG_IRA.pairwiseDeletion.xyiStats,
  CAS_AFG_IRA.globalDeletion.xyiStats)
}
\seealso{
\code{\link[distIUPAC]{xyiStats}},
\code{\link[distIUPAC]{distIUPAC}}, \code{\link[distIUPAC]{rcpp_distIUPAC}}
}
\author{
Kristian K Ullrich
}
