% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.xyStats.R
\name{dist.xyStats}
\alias{dist.xyStats}
\title{dist.xyStats}
\usage{
dist.xyStats(tmpSEQ, x.pos, y.pos, x.name = "x", y.name = "y",
  dist = "IUPAC", ncores = 1)
}
\arguments{
\item{tmpSEQ}{\code{DNAStringSet} [mandatory]}

\item{x.pos}{population X positions [mandatory]}

\item{y.pos}{population Y positions [mandatory]}

\item{x.name}{population X name [default: "x"]}

\item{y.name}{population Y name [default: "y"]}

\item{dist}{distance to use, choose one model as in
\link[ape]{dist.dna} or [default: "IUPAC"]}

\item{ncores}{number of parallel cores to process pairwise distance
calculation [default: 1]}
}
\description{
This function calculates \code{distIUPAC} based distances
comparing two populations (x: receiver; y: donor).
}
\examples{
##Use the 'xyStats' function to handle population assignment automatically

##load sequence data
data("MySequences", package="distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87

##Here, one needs to consider the changed x and y positions due
##to sub-sampling.
CAS.pos_<-seq_along(CAS.pos)
AFG.pos_<-seq(from=length(CAS.pos_)+1, to=length(c(CAS.pos,AFG.pos)))

##pairwise deletion
CAS_AFG.pairwiseDeletion.dist<-distIUPAC(
  as.character(MySequences[c(CAS.pos,AFG.pos)]))
CAS_AFG.pairwiseDeletion.xyStats<-dist.xyStats(
  CAS_AFG.pairwiseDeletion.dist, 
  x.pos=CAS.pos_, y.pos=AFG.pos_,
  x.name="CAS", y.name="AFG")

##global deletion
CAS_AFG.globalDeletion.dist<-distIUPAC(
  as.character(globalDeletion(MySequences[c(CAS.pos,AFG.pos)])))
  CAS_AFG.globalDeletion.xyStats<-dist.xyStats(CAS_AFG.globalDeletion.dist,
  x.pos=CAS.pos_, y.pos=AFG.pos_, x.name="CAS", y.name="AFG")

##compare results
rbind(CAS_AFG.pairwiseDeletion.xyStats, CAS_AFG.globalDeletion.xyStats)
}
\references{
Slatkin, M. (1991). Inbreeding coefficients and coalescence 
times. \emph{Genetics Research}, \bold{58(2)}, 167-175.

Beerli, P. (1998). Structured Populations. \emph{Advances in molecular
ecology}, \bold{306}, 39.

Reich, D., Thangaraj, K., Patterson, N., Price, A. L., & Singh,
L. (2009). Reconstructing Indian population history. \emph{Nature},
\bold{461(7263)}, 489.

Patterson, N., Moorjani, P., Luo, Y., Mallick, S., Rohland, N., Zhan, Y.,
... & Reich, D. (2012). Ancient admixture in human history. \emph{Genetics},
\bold{192(3)}, 1065-1093.

Peter, B. M. (2016). Admixture, population structure, and F-statistics.
\emph{Genetics}, \bold{202(4)}, 1485-1501.

Paradis, E., & Schliep, K. (2018). ape 5.0: an environment for modern
phylogenetics and evolutionary analyses in R. \emph{Bioinformatics},
\bold{35(3)}, 526-528.
}
\seealso{
\code{\link[distIUPAC]{xyStats}},
\code{\link[distIUPAC]{distIUPAC}},
\code{\link[distIUPAC]{rcpp_distIUPAC}},
\code{\link[ape]{dist.dna}}
}
\author{
Kristian K Ullrich
}
