% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin2dnastring.R
\name{dnabin2dnastring}
\alias{dnabin2dnastring}
\title{dnabin2dnastring}
\usage{
dnabin2dnastring(dnabin)
}
\arguments{
\item{dnabin}{\code{DNAbin}}
}
\description{
This function transforms a \code{DNAbin} object from the
\code{ape} package into an \code{DNAStringSet} class of the
\code{Biostrings} package.
}
\examples{
data("woodmouse", package="ape")
alignment<-dnabin2dnastring(woodmouse)
}
\seealso{
\link[Biostrings]{DNAStringSet}, \link[ape]{DNAbin}
}
\author{
Kristian K Ullrich
}
