% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFasta4.R
\name{doFasta4}
\alias{doFasta4}
\title{doFasta4}
\usage{
doFasta4(bam, bam.name = "ind1",
  angsdoption = "-maxDepth 99999 -minQ 20 -minMapQ 30 -iupacRatio 0.2",
  region = NULL, angsd = "angsd", samtools = "samtools",
  system.verbose = FALSE)
}
\arguments{
\item{bam}{indexed bam file which needs to be indexed by 'samtools index'}

\item{bam.name}{individual name [default: "ind1"]}

\item{angsdoption}{angsd option specified as follows
\href{angsd Filters}{http://www.popgen.dk/angsd/index.php/Filters},
\href{Allele Counts}{http://www.popgen.dk/angsd/index.php/Allele_Counts}
and \href{angsd doFasta}{http://www.popgen.dk/angsd/index.php/Fasta}
[default: "-maxDepth 99999 -minQ 20 -minMapQ 30 -iupacRatio 0.2"]}

\item{region}{region specification ('chr' entire chr; 'chr:start'
region from start to end of chr; 'chr:start-stop'
region from start to stop from chr) [default: NULL]}

\item{angsd}{path to angsd executable [default: "angsd"]}

\item{samtools}{path to samtools executable [default: "samtools"]}

\item{system.verbose}{system verbosity output [default: FALSE]}
}
\description{
This function uses the \code{angsd -doFasta 4} option to
obtain IUPAC coded fasta files from an indexed bam file and returns a
\code{DNAStringSet}
}
\examples{
bam.url<-"http://cdna.eva.mpg.de/neandertal/altai/ModernHumans/bam/"
bam.ind1<-"SS6004467-dedup.rg.bam"
bam.file<-paste0(bam.url, bam.ind1)
##region from 5001 to 15000 from chr1 (1:5001-15000)
#bam.region.iupac<-doFasta4(bam=bam.file,
#  angsdoption="-maxDepth 99999 -minQ 20 -minMapQ 30 -iupacRatio 0.2",
#  region="1:5001-15000", angsd="angsd", samtools="samtools")
}
\author{
Kristian K Ullrich
}
