% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrees.R
\name{getTrees}
\alias{getTrees}
\title{getTrees}
\usage{
getTrees(dna, x.pos = NULL, r.pos = 1, wlen = 25000, wjump = 25000,
  wtype = "bp", dist = "IUPAC", model = "bionjs", threads = 1,
  x.name = "x", r.name = "r", chr.name = "chr", plot = FALSE)
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{x.pos}{population X positions}

\item{r.pos}{root position, needs to be within x.pos}

\item{wlen}{sliding windows length}

\item{wjump}{sliding windows jump}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites} or \code{triSites}}

\item{dist}{distance to use}

\item{model}{tree model to use either \link[ape]{njs} or \link[ape]{bionjs}}

\item{threads}{number of parallel threads}

\item{x.name}{population X name}

\item{r.name}{root name}

\item{chr.name}{chromosome name}

\item{plot}{indicates if tree should be plotted}
}
\description{
This function returns nieghbor-joining (\link[ape]{njs} or \link[ape]{bionjs})
trees based on \code{distIUPAC} based distances rooted to a specific indibidual.
}
\examples{
data("MySequences", package = "distIUPAC")
x.pos<-c(1,2,3,4,5,35,63,71,79,82,88,89,97,105,106,114,115,116,117)
r.pos<-116
trees<-getTrees(MySequences, x.pos = x.pos, r.pos = r.pos,
wlen = 10000, wjump = 10000, x.name = "all", r.name = "Rnor", threads = 1)
trees
trees<-getTrees(MySequences, x.pos = x.pos, r.pos = r.pos,
wlen = 10000, wjump = 10000, x.name = "all", r.name = "Rnor", threads = 1, plot = TRUE)
}
\author{
Kristian K Ullrich
}
