% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupac2diploid.R
\name{iupac2diploid}
\alias{iupac2diploid}
\title{iupac2diploid}
\usage{
iupac2diploid(dna)
}
\arguments{
\item{dna}{\code{DNAStringSet} [mandatory]}
}
\description{
This function returns a \code{DNAStringSet} sequence based on a
\code{iupacString} diploid \code{Biostrings} object by splitting each diploid
sequence into two haploid sequences (no random sampling per site, just the
extreme haploid possibilities are used)
}
\examples{
data("MySequences", package = "distIUPAC")
iupac2diploid(MySequences[1])
}
\seealso{
\code{\link[distIUPAC]{triSites}},
\code{\link[distIUPAC]{biSites}}
}
\author{
Kristian K Ullrich
}
