% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupac2gphocs.R
\name{iupac2gphocs}
\alias{iupac2gphocs}
\title{iupac2gphocs}
\usage{
iupac2gphocs(dna, chr = "1", wlen = 25000, start.by = NULL,
  end.by = NULL, threads = 1, pB = TRUE)
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{chr}{set chromosome name}

\item{wlen}{sliding window length for loci retrieval}

\item{start.by}{optional start position}

\item{end.by}{optional end position}

\item{threads}{number of parallel threads}

\item{pB}{specifies if progress should be shown as a progress bar}
}
\description{
This function returns loci blocks as G-PhoCS-like output.
}
\examples{
data("MySequences", package = "distIUPAC")
gphocs<-iupac2gphocs(MySequences, threads = 2)
str(gphocs)
#sink("seqs-sample.txt")
#cat(gphocs,sep="\\n")
#sink(NULL)
}
\author{
Kristian K Ullrich
}
