% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupac2treemix.r
\name{iupac2treemix}
\alias{iupac2treemix}
\title{iupac2treemix}
\usage{
iupac2treemix(dna, list.pos, wlen = 25000, start.by = NULL, end.by = NULL,
  threads = 1, pB = TRUE)
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{list.pos}{population positions list}

\item{wlen}{sliding window length for biSites retrieval}

\item{start.by}{optional start position}

\item{end.by}{optional end position}

\item{threads}{number of parallel threads}

\item{pB}{specifies if progress should be shown as a progress bar}
}
\description{
This function returns biallelic sites as treemix-like output.
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
SPRE.pos<-106:113
pop.list<-list(CAS.pos, AFG.pos, SPRE.pos)
names(pop.list)<-c("CAS", "AFG", "SPRE")
CAS.AFG.SPRE.treemix<-iupac2treemix(MySequences, list.pos = pop.list, threads = 2)
head(CAS.AFG.SPRE.treemix)
}
\author{
Kristian K Ullrich
}
