% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupacString.R
\name{iupacString}
\alias{iupacString}
\title{iupacString}
\usage{
iupacString(dna, name = "iupacString", wlen = 25000, start.by = 1,
  end.by = NULL, threads = 1, pB = TRUE)
}
\arguments{
\item{dna}{\code{DNAStringSet} of length 2 [mandatory]}

\item{wlen}{sliding window length [default: 25000]}

\item{start.by}{optional start position [default: 1]}

\item{end.by}{optional end position [default: NULL]}

\item{threads}{number of parallel threads [default: 1]}

\item{pB}{specifies if progress should be shown as a progress bar
[default: TRUE]}

\item{x.pos}{population X positions [default: NULL]}

\item{min.ind}{minimum number of individuals without gaps ("-", "+", ".")
or without missing sites ("N"), set to size of population ("length(x.pos)")
to mask global deletion sites [default: 0]}
}
\description{
This function returns a \code{IUPAC} sequence based on a
\code{Biostrings} objects
}
\seealso{
\code{\link[distIUPAC]{triSites}},
\code{\link[distIUPAC]{biSites}}
}
\author{
Kristian K Ullrich
}
