% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MySequences-data.R
\docType{data}
\name{MySequences-data}
\alias{MySequences-data}
\alias{MySequences}
\title{MySequences-data}
\format{an object of class \code{DNAStringSet}
see \code{\link[Biostrings]{DNAStringSet}}}
\usage{
data(MySequences)
}
\description{
Example sequence alignment of a region from chromosome 19 of
different mouse species as \code{DNAStringSet}.
}
\examples{
data("MySequences", package="distIUPAC")
}
\keyword{datasets}
