% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posgen.R
\name{posgen}
\alias{posgen}
\title{posgen}
\usage{
posgen(pos, wlen = 50, start.by = NULL, end.by = NULL)
}
\arguments{
\item{pos}{position vector}

\item{wlen}{number of positions per window}

\item{start.by}{optional start position}

\item{end.by}{optional end position}
}
\description{
This function constructs a sliding window matrix based on a SNP position vector
}
\examples{
sw <- posgen(seq(1, 975, by = 5), 50, 1, 1000)
}
\seealso{
\link[distIUPAC]{swgen}, \link[distIUPAC]{biSites}, \link[distIUPAC]{triSites}
}
\author{
Kristian K Ullrich
}
