% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{scoreMatrix}
\alias{scoreMatrix}
\title{scoreMatrix}
\usage{
scoreMatrix()
}
\description{
This function creates a \code{scoreMatrix} object to be used
with the \code{distIUPACmatrix} function.
}
\examples{
data("MySequences", package="distIUPAC")
MyScoreMatrix <- scoreMatrix()
MyScoreMatrix["A","R"] <- 10.0
distIUPACmatrix(as.character(MySequences[1:10]), MyScoreMatrix)
MyScoreMatrix["A","R"] <- 0.5
distIUPACmatrix(as.character(MySequences[1:10]), MyScoreMatrix)
}
\seealso{
\link[distIUPAC]{distIUPACmatrix}, \link[ape]{dist.dna}
}
\author{
Kristian K Ullrich
}
