% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subString.R
\name{subString}
\alias{subString}
\title{subString}
\usage{
subString(x, s, e)
}
\arguments{
\item{x}{\code{DNAStringSet}, \code{RNAString}, \code{AAString}, \code{BString}, \code{DNAStringSet}, \code{RNAStringSet}, \code{AAStringSet}, \code{BStringSet}}

\item{s}{start vector}

\item{e}{end vector}
}
\description{
This function gets a subsequence from a \code{DNAString}, \code{RNAString}, \code{AAString}, \code{BString}, \code{DNAStringSet}, \code{RNAStringSet}, \code{AAStringSet}, \code{BStringSet} object from the \code{Biostrings} package.
}
\examples{
data("MySequences", package = "distIUPAC")
myStart<-c(1,3)
myEnd<-c(2,4)
mySubSequence <- subString(MySequence$dna, myStart, myEnd)
}
\seealso{
\link[Biostrings]{subseq}
}
\author{
Kristian K Ullrich
}
