% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xStats.R
\name{xStats}
\alias{xStats}
\title{xStats}
\usage{
xStats(dna, x.pos = NULL, wlen = 25000, wjump = 25000,
  start.by = NULL, end.by = NULL, wtype = "bp", dist = "IUPAC",
  global.deletion = TRUE, threads = 1, x.name = "x",
  chr.name = "chr", pB = TRUE)
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{x.pos}{population X positions}

\item{wlen}{sliding windows length}

\item{wjump}{sliding windows jump}

\item{start.by}{optional start position}

\item{end.by}{optional end position}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites} or \code{triSites}}

\item{dist}{distance to use}

\item{global.deletion}{a logical indicating whether to delete the sites with missing data in a global or pairwise way (default is to delete in a global way)}

\item{threads}{number of parallel threads}

\item{x.name}{population X name}

\item{chr.name}{chromosome name}

\item{pB}{specifies if progress should be shown as a progress bar}
}
\description{
This function calculates \code{distIUPAC} based distances within one population
(x: receiver; x: donor).
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
CAS.xStats<-xStats(MySequences, x.pos = CAS.pos, x.name = "CAS", threads = 2)
CAS.xStats
}
\author{
Kristian K Ullrich
}
