% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyioStats.R
\name{xyioStats}
\alias{xyioStats}
\title{xyioStats}
\usage{
xyioStats(dna, x.pos, y.pos, i.pos, o.pos, x.name = "x", y.name = "y",
  i.name = "i", o.name = "o", chr.name = "chr", wlen = 25000,
  wjump = 25000, start.by = 1, end.by = NULL, wtype = "bp",
  dist = "IUPAC", global.deletion = TRUE, threads = 1, ncores = 1,
  pB = FALSE, do.ABBA = FALSE, x.freq = 1, y.freq = 1,
  i.freq = 1, o.freq = 1)
}
\arguments{
\item{dna}{\code{DNAStringSet} [mandatory]}

\item{x.pos}{population X positions [P2 population in the four-taxon
scenario (((P1,P2),P3),O) with geneflow from P3>>P2] [mandatory]}

\item{y.pos}{population Y positions [P3 population in the four-taxon
scenario (((P1,P2),P3),O) with geneflow from P3>>P2] [mandatory]}

\item{i.pos}{population I positions [P1 population in the four-taxon
scenario (((P1,P2),P3),O) with geneflow from P3>>P2] [mandatory]}

\item{o.pos}{population I positions [P4 population in the four-taxon
scenario (((P1,P2),P3),O) with geneflow from P3>>P2] [mandatory]}

\item{x.name}{population X name [default: "x"]}

\item{y.name}{population Y name [default: "y"]}

\item{i.name}{population I name [default: "i"]}

\item{o.name}{population O name [default: "o"]}

\item{chr.name}{chromosome name  [default: "chr"]}

\item{wlen}{sliding windows length [default: 25000]}

\item{wjump}{sliding windows jump [default: 25000]}

\item{start.by}{optional start position [default: 1]}

\item{end.by}{optional end position [default: NULL]}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites}
or \code{triSites} [default: "bp"]}

\item{dist}{distance to use [default: "IUPAC"]}

\item{global.deletion}{a logical indicating whether to delete the sites
with missing data in a global or pairwise way (default is to delete in a
global way) [default: TRUE]}

\item{threads}{number of parallel threads [default: 1]}

\item{ncores}{number of parallel cores to process pairwise distance
calculation [default: 1] see \code{\link[distIUPAC]{rcpp_distIUPAC}}}

\item{pB}{specifies if progress should be shown as a progress bar
[default: FALSE]}

\item{do.ABBA}{specifies if ABBA-BABA statistics should be calculated
[default: FALSE]}

\item{x.freq}{[default: 1.0]}

\item{y.freq}{[default: 1.0]}

\item{i.freq}{[default: 1.0]}

\item{o.freq}{[default: 1.0]}
}
\description{
This function calculates \code{distIUPAC} based distances
comparing two populations (x: receiver and y: donor) with an ingroup
population (i: ingroup) and an outgroup population (o: outgroup).
In the four-taxon scenario (((P1,P2),P3),O) with geneflow from P3>>P2,
the populations should be defined for deltaMean and deltaMin statistics as
follows [x:P2 y:P3 i:P1 o:P4].
Accordingly in the four-taxon scenario (((P1,P2),P3),O) with geneflow from
P2>>P3, the populations should be defined for RND, Gmin and RNDmin
statistics as follows [x:P3 y:P2 i:P1 o:P4]. Optional, ABBA-BABA statisctis
can be calculated concordantly on the four-taxon scenario
(((P1:i,P2:x),P3:y),O:o).
}
\examples{
data("MySequences", package="distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
SPRE.pos<-106:113
APO.pos<-1
AFG.SPRE.CAS.APO.xyioStats<-xyioStats(MySequences, x.pos=AFG.pos,
y.pos=SPRE.pos, i.pos=CAS.pos, o.pos=APO.pos,
x.name="AFG", y.name="SPRE", i.name="CAS", o.name="APO",
threads=2)
AFG.SPRE.CAS.APO.xyioStats
}
\author{
Kristian K Ullrich
}
