% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyStats.R
\name{xyStats}
\alias{xyStats}
\title{xyStats}
\usage{
xyStats(dna, x.pos, y.pos, x.name = "x", y.name = "y",
  chr.name = "chr", wlen = 25000, wjump = 25000, start.by = 1,
  end.by = NULL, wtype = "bp", dist = "IUPAC",
  global.deletion = TRUE, threads = 1, ncores = 1, pB = FALSE)
}
\arguments{
\item{dna}{\code{DNAStringSet} [mandatory]}

\item{x.pos}{population X positions [mandatory]}

\item{y.pos}{population Y positions [mandatory]}

\item{x.name}{population X name [default: "x"]}

\item{y.name}{population Y name [default: "y"]}

\item{chr.name}{chromosome name [default: "chr"]}

\item{wlen}{sliding windows length [default: 25000]}

\item{wjump}{sliding windows jump [default: 25000]}

\item{start.by}{optional start position [default: 1]}

\item{end.by}{optional end position [default: NULL]}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites}
or \code{triSites} [default: "bp"]}

\item{dist}{distance to use, choose one model as in
\link[ape]{dist.dna} or [default: "IUPAC"]}

\item{global.deletion}{a logical indicating whether to delete the sites
with missing data in a global or pairwise way (default is to delete in a
global way) [default: TRUE]}

\item{threads}{number of parallel threads [default: 1]}

\item{ncores}{number of parallel cores to process pairwise distance
calculation [default: 1]}

\item{pB}{specifies if progress should be shown as a progress bar
[default: FALSE]}
}
\description{
This function calculates \code{distIUPAC} based distances
comparing two populations (x: receiver and y: donor).
}
\examples{
##load sequence data
data("MySequences", package="distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
##sliding windows based on base-pair length
CAS.AFG.xyStats<-xyStats(MySequences, x.pos=CAS.pos, y.pos=AFG.pos,
x.name="CAS", y.name="AFG", threads=2)
CAS.AFG.xyStats
##sliding windows based on biSites
CAS.AFG.xyStats<-xyStats(MySequences, x.pos=CAS.pos, y.pos=AFG.pos,
wlen=50, wtype="biSites", x.name="CAS", y.name="AFG", threads=2)
CAS.AFG.xyStats
}
\references{
Slatkin, M. (1991). Inbreeding coefficients and coalescence 
times. \emph{Genetics Research}, \bold{58(2)}, 167-175.

Beerli, P. (1998). Structured Populations. \emph{Advances in molecular
ecology}, \bold{306}, 39.

Reich, D., Thangaraj, K., Patterson, N., Price, A. L., & Singh,
L. (2009). Reconstructing Indian population history. \emph{Nature},
\bold{461(7263)}, 489.

Patterson, N., Moorjani, P., Luo, Y., Mallick, S., Rohland, N., Zhan, Y.,
... & Reich, D. (2012). Ancient admixture in human history. \emph{Genetics},
\bold{192(3)}, 1065-1093.

Peter, B. M. (2016). Admixture, population structure, and F-statistics.
\emph{Genetics}, \bold{202(4)}, 1485-1501.

Paradis, E., & Schliep, K. (2018). ape 5.0: an environment for modern
phylogenetics and evolutionary analyses in R. \emph{Bioinformatics},
\bold{35(3)}, 526-528.
}
\seealso{
\code{\link[distIUPAC]{dist.xyStats}},
\code{\link[ape]{dist.dna}}
}
\author{
Kristian K Ullrich
}
