% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.xStats.R
\name{dist.xStats}
\alias{dist.xStats}
\title{dist.xStats}
\usage{
dist.xStats(dIUPAC, x.pos = NULL, x.name = "x")
}
\arguments{
\item{dIUPAC}{\code{distIUPAC} distance matrix including \code{sitesUsed} matrix}

\item{x.pos}{population X positions}

\item{x.name}{population X name}
}
\description{
This function calculates \code{distIUPAC} based distances within one population
(x: receiver; x: donor).
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
#pairwise deletion
CAS.pairwiseDeletion.dist<-distIUPAC(as.character(MySequences[CAS.pos]))
CAS.pairwiseDeletion.xStats<-dist.xStats(CAS.pairwiseDeletion.dist, x.name="CAS")
#global deletion
CAS.globalDeletion.dist<-distIUPAC(as.character(globalDeletion(MySequences[CAS.pos])))
CAS.globalDeletion.xStats<-dist.xStats(CAS.globalDeletion.dist, x.name="CAS")
#compare results
rbind(CAS.pairwiseDeletion.xStats, CAS.globalDeletion.xStats)
}
\author{
Kristian K Ullrich
}
