% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distIUPACsw.R
\name{distIUPACsw}
\alias{distIUPACsw}
\title{distIUPACsw}
\usage{
distIUPACsw(dna, FUN = NULL, chr.name = "chr", wlen = 25000,
  wjump = 25000, start.by = NULL, end.by = NULL, wtype = "bp",
  dist = "IUPAC", global.deletion = TRUE, threads = 1, pB = TRUE)
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{FUN}{the function to be applied}

\item{wlen}{sliding windows length}

\item{wjump}{sliding windows jump}

\item{start.by}{optional start position}

\item{end.by}{optional end position}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites} or \code{triSites}}

\item{dist}{distance to use}

\item{global.deletion}{a logical indicating whether to delete the sites with missing data in a global or pairwise way (default is to delete in a global way)}

\item{threads}{number of parallel threads}

\item{pB}{specifies if progress should be shown as a progress bar}
}
\description{
This function calculates \code{IUPAC} distances on \code{Biostrings} \code{DNAStringSet} objects
on sliding windows which can be used to parse to further functions
}
\examples{
data("MySequences", package = "distIUPAC")
}
\author{
Kristian K Ullrich
}
