% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnastring2apealg.R
\name{dnastring2apealg}
\alias{dnastring2apealg}
\title{dnastring2aperalg}
\usage{
dnastring2apealg(dna)
}
\arguments{
\item{dna}{\code{DNAStringSet}}
}
\description{
This function transforms a \code{DNAStringSet} object from the \code{Biostrings} package into an \code{alignment} class of the \code{ape} package.
}
\examples{
data("MySequences", package = "distIUPAC")
alignment <- dnastring2apealg(MySequences)
alignment.bin <- as.DNAbin(alignment)
}
\seealso{
\link[Biostrings]{DNAStringSet}, \link[ape]{as.alignment}
}
\author{
Kristian K Ullrich
}
