% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFasta2.R
\name{doFasta2}
\alias{doFasta2}
\title{doFasta2}
\usage{
doFasta2(bamlist, bamlist.name = "pop1",
  angsdoption = "-maxDepth 99999 -minQ 20 -minMapQ 30", region = NULL,
  angsd = "angsd", samtools = "samtools", system.verbose = FALSE)
}
\arguments{
\item{bamlist}{list of indexed bam files which needs to be indexed by 'samtools index'}

\item{bamlist.name}{population name}

\item{angsdoption}{angsd option specified as here \href{angsd Filters}{http://www.popgen.dk/angsd/index.php/Filters},
here \href{angsd Allele Counts}{http://www.popgen.dk/angsd/index.php/Allele_Counts} and here \href{angsd doFasta}{http://www.popgen.dk/angsd/index.php/Fasta}}

\item{region}{region specification ('chr' entire chr; 'chr:start' region from start to end of chr; 'chr:start-stop' region from start to stop from chr)}

\item{angsd}{path to angsd executable}

\item{samtools}{path to samtools executable}

\item{system.verbose}{system verbosity output}
}
\description{
This function uses the \code{angsd -doFasta 2} option to obtain a consensus
fasta file from an bam file list of indexed bam files and returns a \code{DNAStringSet}
}
\examples{
bam.list<-c(system.file("extdata", "ind1.pop1.bam", package="distIUPAC"),system.file("extdata", "ind2.pop1.bam", package="distIUPAC"))
#complete 
bam.consensus<-doFasta2(bamlist=bam.list, angsdoption="-maxDepth 99999 -minQ 20 -minMapQ 30", angsd="angsd", samtools="samtools")
#region from 5001 to 15000 from chr1 (chr1:5001-15000)
bam.region.consensus<-doFasta2(bamlist=bam.list, angsdoption="-maxDepth 99999 -minQ 20 -minMapQ 30", region="chr1:5001-15000", angsd="angsd", samtools="samtools")
}
\author{
Kristian K Ullrich
}
