% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFasta4.R
\name{doFasta4}
\alias{doFasta4}
\title{doFasta4}
\usage{
doFasta4(bam, bam.name = "ind1",
  angsdoption = "-maxDepth 99999 -minQ 20 -minMapQ 30 -iupacRatio 0.2",
  region = NULL, angsd = "angsd", samtools = "samtools",
  system.verbose = FALSE)
}
\arguments{
\item{bam}{indexed bam file which needs to be indexed by 'samtools index'}

\item{bam.name}{individual name}

\item{angsdoption}{angsd option specified as here \href{angsd Filters}{http://www.popgen.dk/angsd/index.php/Filters},
here \href{angsd Allele Counts}{http://www.popgen.dk/angsd/index.php/Allele_Counts} and here \href{angsd doFasta}{http://www.popgen.dk/angsd/index.php/Fasta}}

\item{region}{region specification ('chr' entire chr; 'chr:start' region from start to end of chr; 'chr:start-stop' region from start to stop from chr)}

\item{angsd}{path to angsd executable}

\item{samtools}{path to samtools executable}

\item{system.verbose}{system verbosity output}
}
\description{
This function uses the \code{angsd -doFasta 4} option to obtain IUPAC coded
fasta files from an indexed bam file and returns a \code{DNAStringSet}
}
\examples{
bam.file<-system.file("extdata", "ind1.pop1.bam", package="distIUPAC")
#complete 
bam.iupac<-doFasta4(bam=bam.file, angsdoption="-maxDepth 99999 -minQ 20 -minMapQ 30 -iupacRatio 0.2", angsd="angsd", samtools="samtools")
#region from 5001 to 15000 from chr1 (chr1:5001-15000)
bam.region.iupac<-doFasta4(bam=bam.file, angsdoption="-maxDepth 99999 -minQ 20 -minMapQ 30 -iupacRatio 0.2", region="chr1:5001-15000", angsd="angsd", samtools="samtools")
}
\author{
Kristian K Ullrich
}
