% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrees.R
\name{getTrees}
\alias{getTrees}
\title{getTrees}
\usage{
getTrees(dna, x.pos = NULL, r.pos = 1, resolve.root = TRUE,
  wlen = 25000, wjump = 25000, start.by = 1, end.by = NULL,
  wtype = "bp", dist = "IUPAC", global.deletion = TRUE,
  model = "bionjs", threads = 1, ncores = 1, x.name = "x",
  r.name = "r", chr.name = "chr", plot = FALSE, pB = FALSE)
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{x.pos}{population X positions [default: NULL]
take all sequences from \code{DNAStringSet}}

\item{r.pos}{root position, needs to be within x.pos [default: 1]}

\item{resolve.root}{adds a zero-length branch below the MRCA
of the ingroup \link[ape]{root} [default: TRUE]}

\item{wlen}{sliding windows length  [default: 25000]}

\item{wjump}{sliding windows jump [default: 25000]}

\item{start.by}{optional start position [default: 1]}

\item{end.by}{optional end position [default: NULL]}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites}
or \code{triSites}  [default: "bp"]}

\item{dist}{distance to use, choose one model as in
\link[ape]{dist.dna} or [default: "IUPAC"]}

\item{global.deletion}{a logical indicating whether to delete the sites
with missing data in a global [default: TRUE] or pairwise way [FALSE]}

\item{model}{tree model to use either \link[ape]{njs}
or \link[ape]{bionjs} [default: "bionjs"]}

\item{threads}{number of parallel threads [default: 1]}

\item{ncores}{number of parallel cores to process pairwise distance
calculation [default: 1]}

\item{x.name}{population X name [default: "x"]}

\item{r.name}{root name [default: "r"]}

\item{chr.name}{chromosome name [default: "chr"]}

\item{plot}{indicates if tree should be plotted [default: FALSE]}

\item{pB}{specifies if progress should be shown as a progress bar
[default: FALSE]}
}
\description{
This function returns nieghbor-joining (\link[ape]{njs} or \link[ape]{bionjs})
trees based on \code{distIUPAC} based distances rooted to a specific individual.
}
\examples{
data("MySequences", package = "distIUPAC")
x.pos<-c(1,2,3,4,5,35,63,71,79,82,88,89,97,105,106,114,115,116,117)
r.pos<-116
trees<-getTrees(MySequences, x.pos = x.pos, r.pos = r.pos,
wlen = 10000, wjump = 10000, x.name = "all", r.name = "Rnor", threads = 1)
trees
trees<-getTrees(MySequences, x.pos = x.pos, r.pos = r.pos,
wlen = 10000, wjump = 10000, x.name = "all", r.name = "Rnor",
threads = 2, plot = TRUE)
}
\author{
Kristian K Ullrich
}
