% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalDeletion.R
\name{globalDeletion}
\alias{globalDeletion}
\title{globalDeletion}
\usage{
globalDeletion(dna)
}
\arguments{
\item{dna}{\code{DNAStringSet}}
}
\description{
This function returns a \code{DNAStringSet} reduced by all sites containing any gaps ("-", "+", ".") or missing ("N") sites
}
\examples{
data("MySequences", package = "distIUPAC")
mySubSequence <- globalDeletion(MySequences)
}
\seealso{
\link[distIUPAC]{pairwiseDeletion}, \link[distIUPAC]{biSites}, \link[distIUPAC]{triSites}
}
\author{
Kristian K Ullrich
}
