% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupacString.R
\name{iupacString}
\alias{iupacString}
\title{iupacString}
\usage{
iupacString(dna, x.pos = 1, y.pos = 2, name = "iupacString",
  wlen = 25000, start.by = 1, end.by = NULL, threads = 1,
  pB = FALSE)
}
\arguments{
\item{dna}{\code{DNAStringSet} of length 2 [mandatory]}

\item{x.pos}{population X positions [default: 1]}

\item{y.pos}{population X positions [default: 2]}

\item{wlen}{sliding window length [default: 25000]}

\item{start.by}{optional start position [default: 1]}

\item{end.by}{optional end position [default: NULL]}

\item{threads}{number of parallel threads [default: 1]}

\item{pB}{specifies if progress should be shown as a progress bar
[default: TRUE]}
}
\description{
This function returns a \code{IUPAC} sequence based on a
\code{Biostrings} objects
}
\examples{
data("MySequences", package = "distIUPAC")
iupacString(MySequences, x.pos=5, y.pos=6)
}
\seealso{
\code{\link[distIUPAC]{triSites}},
\code{\link[distIUPAC]{biSites}}
}
\author{
Kristian K Ullrich
}
