\name{pairwiseDeletion}
\alias{pairwiseDeletion}
\title{pairwiseDeletion}
\usage{
pairwiseDeletion()
}
\description{
}
\examples{
data("woodmouse", , package = "ape")
pairwiseDeletion(as.character(dnabin2dnastring(woodmouse)))
}
\author{
Kristian K Ullrich
}
\seealso{
\link[distIUPAC]{distIUPAC}, \link[distIUPAC]{distIUPACmatrix}, \link[ape]{dist.dna}
}
