% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{scoreMatrix}
\alias{scoreMatrix}
\title{scoreMatrix}
\usage{
scoreMatrix()
}
\description{
This function creates a \code{scoreMatrix} object to be used
with the \code{distIUPACmatrix} function. By default, the score matrix is
defined as literal distance obtained from \code{Chang et al. 2017}.
(see \url{https://link.springer.com/article/10.1007/s00335-017-9704-9})
}
\examples{
data("MySequences", package="distIUPAC")
MyScoreMatrix <- scoreMatrix()
MyScoreMatrix["A","R"] <- 10.0
distIUPACmatrix(as.character(MySequences[1:10]), MyScoreMatrix)
MyScoreMatrix["A","R"] <- 0.5
distIUPACmatrix(as.character(MySequences[1:10]), MyScoreMatrix)
}
\references{
Chang, P. L., Kopania, E., Keeble, S., Sarver, B. A., Larson,
E., Orth, A., ... & Dean, M. D. (2017). Whole exome sequencing of
wild-derived inbred strains of mice improves power to link phenotype and
genotype. \emph{Mammalian genome}, \bold{28(9-10)}, 416-425.
}
\seealso{
\link[distIUPAC]{distIUPACmatrix}, \link[ape]{dist.dna}
}
\author{
Kristian K Ullrich
}
