% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{scoreMatrix}
\alias{scoreMatrix}
\title{scoreMatrix}
\usage{
scoreMatrix()
}
\description{
This function creates a \code{scoreMatrix} object to be used with the \code{distIUPACmatrix} function.
}
\examples{
MyScoreMatrix <- scoreMatrix()
MyScoreMatrix["A","R"] <- 0.5
data("MySequences", package = "distIUPAC")
distIUPACmatrix(as.character(MySequences), MyScoreMatrix)
}
\seealso{
\link[distIUPAC]{distIUPACmatrix}, \link[ape]{dist.dna}
}
\author{
Kristian K Ullrich
}
