% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swgen.R
\name{swgen}
\alias{swgen}
\title{swgen}
\usage{
swgen(wlen = 1000, wjump = 100, start.by = 1, end.by = 1000)
}
\arguments{
\item{wlen}{window size}

\item{wjump}{window jump}

\item{start.by}{start position}

\item{end.by}{end position}
}
\description{
This function constructs a sliding window matrix
}
\examples{
sw <- swgen(100, 100, 1, 1000)
}
\seealso{
\link[distIUPAC]{posgen}
}
\author{
Kristian K Ullrich
}
