% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyMultiStats.R
\name{xyMultiStats}
\alias{xyMultiStats}
\title{xyMultiStats}
\usage{
xyMultiStats(dna, list.pos, wlen = 25000, wjump = 25000, start.by = NULL,
  end.by = NULL, wtype = "bp", dist = "IUPAC", threads = 1,
  chr.name = "chr")
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{list.pos}{population positions list}

\item{wlen}{sliding windows length}

\item{wjump}{sliding windows jump}

\item{start.by}{optional start position}

\item{end.by}{optional end position}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites} or \code{triSites}}

\item{dist}{distance to use}

\item{threads}{number of parallel threads}

\item{chr.name}{chromosome name}
}
\description{
This function calculates \code{distIUPAC} based distances comparing all
possible pairwise population combinations.
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
SPRE.pos<-106:113
pop.list<-list(CAS.pos, AFG.pos, SPRE.pos)
names(pop.list)<-c("CAS", "AFG", "SPRE")
#sliding windows based on base-pair length
CAS.AFG.SPRE.xyMultiStats<-xyMultiStats(MySequences, list.pos = pop.list, threads = 2)
CAS.AFG.SPRE.xyMultiStats
#sliding windows based on biSites
CAS.AFG.SPRE.xyMultiStats<-xyMultiStats(MySequences, list.pos = pop.list,
wtype = "biSites", wlen = 50, threads = 2)
CAS.AFG.SPRE.xyMultiStats
}
\author{
Kristian K Ullrich
}
