% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyoStats.R
\name{xyoStats}
\alias{xyoStats}
\title{xyoStats}
\usage{
xyoStats(dna, x.pos, y.pos, o.pos, wlen = 25000, wjump = 25000,
  start.by = NULL, end.by = NULL, wtype = "bp", dist = "IUPAC",
  global.deletion = FALSE, threads = 1, x.name = "x", y.name = "y",
  o.name = "o", chr.name = "chr")
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{x.pos}{population X positions}

\item{y.pos}{population Y positions}

\item{o.pos}{population O positions}

\item{wlen}{sliding windows length}

\item{wjump}{sliding windows jump}

\item{start.by}{optional start position}

\item{end.by}{optional end position}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites} or \code{triSites}}

\item{dist}{distance to use}

\item{global.deletion}{a logical indicating whether to delete the sites with missing data in a global way (default is to delete in a pairwise way)}

\item{threads}{number of parallel threads}

\item{x.name}{population X name}

\item{y.name}{population Y name}

\item{o.name}{population O name}

\item{chr.name}{chromosome name}
}
\description{
This function calculates \code{distIUPAC} based distances comparing two populations
(x: receiver; y: donor) with an outgroup population (o: outgroup).
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
SPRE.pos<-106:113
AFG.SPRE.CAS.xyoStats<-xyoStats(MySequences, x.pos = AFG.pos, y.pos = SPRE.pos, o.pos = CAS.pos,
threads = 2, x.name = "AFG", y.name = "SPRE", o.name = "CAS")
AFG.SPRE.CAS.xyoStats
}
\author{
Kristian K Ullrich
}
