% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyStats.R
\name{xyStats}
\alias{xyStats}
\title{xyStats}
\usage{
xyStats(dna, x.pos, y.pos, wlen = 25000, wjump = 25000, wtype = "bp",
  dist = "IUPAC", threads = 1, x.name = "x", y.name = "y",
  chr.name = "chr")
}
\arguments{
\item{dna}{\code{DNAStringSet}}

\item{x.pos}{population X positions}

\item{y.pos}{population Y positions}

\item{wlen}{sliding windows length}

\item{wjump}{sliding windows jump}

\item{wtype}{sliding windows type to use \code{bp}, \code{biSites} or \code{triSites}}

\item{dist}{distance to use}

\item{threads}{number of parallel threads}

\item{x.name}{population X name}

\item{y.name}{population Y name}

\item{chr.name}{chromosome name}
}
\description{
This function calculates \code{distIUPAC} based distances comparing two populations 
(x: receiver; y: donor).
}
\examples{
data("MySequences", package = "distIUPAC")
CAS.pos<-5:34
AFG.pos<-82:87
#sliding windows based on base-pair length
CAS.AFG.xyStats<-xyStats(MySequences, x.pos = CAS.pos, y.pos = AFG.pos,
x.name = "CAS", y.name = "AFG", threads = 4)
CAS.AFG.xyStats
#sliding windows based on biSites
CAS.AFG.xyStats<-xyStats(MySequences, x.pos = CAS.pos, y.pos = AFG.pos,
wlen = 50, wtype = "biSites", x.name = "CAS", y.name = "AFG", threads = 4)
CAS.AFG.xyStats
}
\author{
Kristian K Ullrich
}
